+++ src-docs/src/views/collapsible_nav/collapsible_nav_all.tsx
@@ -300,7 +300,6 @@ const CollapsibleNavAll = () => {
         sections={[
           {
             items: leftSectionItems,
-            borders: 'right',
           },
           {
             items: [