+++ src/components/code/code_block_annotations.test.tsx
@@ -8,7 +8,7 @@
 
 import React from 'react';
 
-import { fireEvent } from '@testing-library/dom';
+import { fireEvent } from '@testing-library/react';
 import { render, waitForEuiPopoverOpen } from '../../test/rtl';
 import { shouldRenderCustomStyles } from '../../test/internal';
 import { requiredProps } from '../../test';