+++ src/components/combo_box/combo_box.test.tsx
@@ -7,7 +7,7 @@
  */
 
 import React, { ReactNode } from 'react';
-import { fireEvent } from '@testing-library/react';
+import { act, fireEvent } from '@testing-library/react';
 import { shallow, mount } from 'enzyme';
 import { render } from '../../test/rtl';
 import {
@@ -91,7 +91,9 @@ describe('props', () => {
       />
     );
 
-    component.setState({ isListOpen: true });
+    act(() => {
+      component.setState({ isListOpen: true });
+    });
     expect(takeMountedSnapshot(component)).toMatchSnapshot();
   });
 
@@ -136,7 +138,10 @@ describe('props', () => {
 
     test('renders in the options dropdown', () => {
       const component = mount(<EuiComboBox options={options} />);
-      component.setState({ isListOpen: true });
+
+      act(() => {
+        component.setState({ isListOpen: true });
+      });
 
       const dropdown = component.find(
         'div[data-test-subj="comboBoxOptionsList"]'
@@ -329,10 +334,16 @@ describe('behavior', () => {
         />
       );
 
-      component.setState({ searchValue: 'foo' });
-      const searchInput = findTestSubject(component, 'comboBoxSearchInput');
-      searchInput.simulate('focus');
-      searchInput.simulate('keyDown', { key: comboBoxKeys.ENTER });
+      act(() => {
+        component.setState({ searchValue: 'foo' });
+      });
+
+      act(() => {
+        const searchInput = findTestSubject(component, 'comboBoxSearchInput');
+        searchInput.simulate('focus');
+        searchInput.simulate('keyDown', { key: comboBoxKeys.ENTER });
+      });
+
       expect(onCreateOptionHandler).toHaveBeenCalledTimes(1);
       expect(onCreateOptionHandler).toHaveBeenNthCalledWith(1, 'foo', options);
     });
@@ -391,15 +402,20 @@ describe('behavior', () => {
         />
       );
 
-      component.setState({ searchValue: 'foo' });
       const searchInput = findTestSubject(component, 'comboBoxSearchInput');
-      searchInput.simulate('focus');
 
-      const searchInputNode = searchInput.getDOMNode();
-      // React doesn't support `focusout` so we have to manually trigger it
-      searchInputNode.dispatchEvent(
-        new FocusEvent('focusout', { bubbles: true })
-      );
+      act(() => {
+        component.setState({ searchValue: 'foo' });
+        searchInput.simulate('focus');
+      });
+
+      act(() => {
+        const searchInputNode = searchInput.getDOMNode();
+        // React doesn't support `focusout` so we have to manually trigger it
+        searchInputNode.dispatchEvent(
+          new FocusEvent('focusout', { bubbles: true })
+        );
+      });
 
       expect(onCreateOptionHandler).toHaveBeenCalledTimes(1);
       expect(onCreateOptionHandler).toHaveBeenNthCalledWith(1, 'foo', options);