+++ src/components/datagrid/controls/column_selector.test.tsx
@@ -7,9 +7,10 @@
  */
 
 import React from 'react';
-import { act } from 'react-dom/test-utils';
+import { act } from '@testing-library/react';
 import { shallow, mount, ReactWrapper } from 'enzyme';
 import { findTestSubject } from '../../../test';
+import { testByReactVersion } from '../../../test/internal';
 
 import { EuiDataGridToolBarVisibilityOptions } from '../data_grid_types';
 
@@ -76,15 +77,18 @@ describe('useDataGridColumnSelector', () => {
       component.setProps({});
     };
 
-    it('renders a toolbar button/popover allowing users to set column visibility and order', () => {
-      const component = mount(<MockComponent showColumnSelector={true} />);
-      openPopover(component);
-      expect(component.render()).toMatchSnapshot();
-      expect(
-        component.find('[data-popover-panel]').first().render()
-      ).toMatchSnapshot();
-      closePopover(component);
-    });
+    testByReactVersion(
+      'renders a toolbar button/popover allowing users to set column visibility and order',
+      () => {
+        const component = mount(<MockComponent showColumnSelector={true} />);
+        openPopover(component);
+        expect(component.render()).toMatchSnapshot();
+        expect(
+          component.find('[data-popover-panel]').first().render()
+        ).toMatchSnapshot();
+        closePopover(component);
+      }
+    );
 
     it('does not render if all valid sub-options are disabled', () => {
       const component = shallow(