+++ src/components/datagrid/controls/keyboard_shortcuts.test.tsx
@@ -7,19 +7,28 @@
  */
 
 import React from 'react';
-import { renderHook, act } from '@testing-library/react-hooks';
-import { render, waitForEuiPopoverOpen } from '../../../test/rtl';
+import {
+  render,
+  renderHook,
+  renderHookAct,
+  waitForEuiPopoverOpen,
+} from '../../../test/rtl';
 import { useDataGridKeyboardShortcuts } from './keyboard_shortcuts';
+import { fireEvent } from '@testing-library/react';
 
 describe('useDataGridKeyboardShortcuts', () => {
   it('returns a popover containing a list of keyboard shortcuts', async () => {
     const { result } = renderHook(() => useDataGridKeyboardShortcuts());
     const { baseElement, getByTestSubject, rerender } = render(
-      <>{result.current.keyboardShortcuts}</>
+      <div data-test-subj="hookRoot">{result.current.keyboardShortcuts}</div>
     );
 
-    act(() => getByTestSubject('dataGridKeyboardShortcutsButton').click());
-    rerender(<>{result.current.keyboardShortcuts}</>);
+    renderHookAct(() => {
+      fireEvent.click(getByTestSubject('dataGridKeyboardShortcutsButton'));
+    });
+    rerender(
+      <div data-test-subj="hookRoot">{result.current.keyboardShortcuts}</div>
+    );
     await waitForEuiPopoverOpen();
 
     expect(baseElement).toMatchSnapshot();