+++ src/components/form/validatable_control/validatable_control.test.tsx
@@ -8,7 +8,7 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '../../../test/rtl';
 
 import {
   EuiValidatableControl,